Add files via upload
The objective of this project is to build a simple Todo List app that allows users to add, edit and delete tasks. And store the tasks in Local storage so that they persist even after the browser is closed.
The Key concept used are: React state, event handling, local Storage and conditional rendering.